Image-Captioning/requirements.txt
absl-py==0.12.0
astor==0.8.1
astunparse==1.6.3
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
comtypes==1.1.9
flatbuffers==1.12
gast==0.3.3
google-auth==1.28.0
google-auth-oauthlib==0.4.3
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
importlib-metadata==3.8.0
joblib==1.0.1
Keras==2.2.0
Keras-Applications==1.0.2
Keras-Preprocessing==1.0.1
Markdown==3.3.4
mock==4.0.3
nltk==3.5
numpy==1.19.5
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.1.5
Pillow==8.1.2
progressbar==2.5
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pypiwin32==223
python-dateutil==2.8.1
pyttsx==1.1
pyttsx3==2.90
pytz==2021.1
pywin32==300
PyYAML==5.4.1
regex==2021.3.17
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
scipy==1.5.4
six==1.15.0
tensorboard==1.13.1
tensorboard-plugin-wit==1.8.0
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
Theano==1.0.5
tqdm==4.59.0
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.4.1
